Modify methods in PlayerChange
Start method initialises maxTotemCount. It also enables the first totem
and disables the other totems. Totem 0 will always be active first.

A list variable, listTotems, contains all totems in the scene.
ChangeCharacter method will disable the current active totem, increment
a counter, and then enable the next totem in the list. When the counter
equals or is larger than maxTotemCount, the counter is reset to 0. This
is so that when the player switches characters, it loops round all of
them.