Add StackTotem class Modify totem prefab
StackTotem class is sets up the empty game objects that will be used
to correctly position totems when they are stacked on top of the active
totem. It will instantiate a total of maxTotemCount-1 empty gameobjects
above the totem at the start of the scene.

The totem prefab now has a gameobject, totemStackPositions, that
will contain all of the gameobjects created by StackTotem.