Add StackTotemInPos to StackTotem Modify PlayerChange Modify PlayerMove
StackTotemInPos sets the to-be-stacked totem to one of the position
gameobjects and sets the to-be-stacked totem's parent to be the active
totem's totemStackPositions container.

PlayerChange now sets the kinematics on stacked totems to be true to
stop the rigidbody physics from affecting the stacked totem. This
keeps the stacked totem above the active totem.

PlayerMove has a isTotemStacked boolean. This is used in a check in
OnCollisionEnter2D to ensure that StackTotemInPos is only called if the
totem it collides into is already stacked.

There is an issue with hump height being dramitcally decreased when
stacking occurs.