Add TotemCollision Class Move OnCollisionEnter2D to TotemCollision Modified PlayerMove Class
There was an issue with PlayerMove.OnCollisionEnter2D causing a null
exception error when a totem that collides into an non-recruited totem
after performing a totem jump. This is due to PlayerMove being disabled
after performing a totem jump.

This was resolved by moving the OnCollisionEnter2D to the new
TotemCollision Class, so that it remains active after the totem jump.