First commit to branch totem-throw Add ThrowTotem class. Add TossTopTotem method.
ThrowTotem class created and added to totem prefab.

TossTopTotem will eventually disconnect the topmost totem in stack and
"throw" it. Currently, it only prints a debug message.

Added a new check in PlayerMove.Update that checks for whether the
"throw" key was pressed and that the StackTotem.CountStackTotems is
greater than 0.