Add ThrowTotem.SetUpThrow Modify StackTotem and PlayerMove
SetUpThrow is called to set up the throwing vector. TossTopTotem now
enables the isKinematic, applies the velocity to the totem, and sets
isTotemStacked to false.

StackTotem.GetNextTotem has had the line ChangeToTargetTotem removed,
as SetUpThrow needs to setup the totem but will not change control to
the thrown totem. ChangeToTargetTotem has been moved to PlayerMove.Jump.