Add new methods into UIEndgame Modify FinishLevel, UIManagerStage, UIEndgame
UIEndgame.SetUpEndgameMessage is now SetEndgameScreen.

UIManagerStage.TriggerEndgame modified to take 2 inputs:
int levelNum, and bool isLastLevel.
These are passed from FinishLevel when calling TriggerEndgame.

These are passed to SetEndgameScreen to set up the endgame overlay.
When isLastLevel is false, both the nextLevel button and mainMenu
button are visible, and the levelNum variable is set to load the
appropriate level. When true, nextLevel button is disabled and
mainMenu button is reposition to the center of the screen.

GoToNextLevel and ReturnToMainLevel have been added to UIEndgame, and
are called when clicking the nextLevel button and mainMenu button
respectively. GoToNextLevel uses levelNum to load the next level.
ReturnToMainLevel loads the mainMenu level.