Add methods to UIPauseMenu Add methods to UIManagerStage Modify PlayerMove
Methods PauseGame and ContinueGame added to UIPauseMenu.
PauseGame sets Time.timeScale to 0. ContinueGame sets
Time.timeScale to 1.

There was an issue where clicking the Unpause button would cause make
pressing the spacebar to pause or unpause the game. This has been
resolved by making PauseGame and ContinueGame enable and disable the
interactivity of the buttons on the pauseMenu overlay.

SetPauseOverlay added to UIManagerStage. A new boolean variable called
isGamePaused has been added. When SetPauseOverlay is called,
isGamePaused will be inverted and will disable/enable the inLevel
overlay and the pauseMenu overlays and call either PauseGame or
ContinueGame

PlayerMove calls SetPauseOverlay when the P key is pressed.