Add CameraSetUp class Add cameraBounds prefab
CameraSetUp class sets up the camera's position and orthographicSize to
so that the entire stage is visible. This requires two reference
points, lowerLeftBound and upperRightBound. These are empty game
objects that are set up in the scene editor.

cameraBounds prefab contains the main camera, lowerLeftBound and
upperRightBound gameobjects, and the CameraSetUp script.