Add levels 2 and 3 Modified PlayerMove.Jump Modified throwVector and throwPower in TotemThrow
There were minor issues with designing levels 2 and 3, as it is
possible to complete the level without using the required ability.
Had to tweak few values that are listed below.

Jump now contains an additional modifier when doing a totem jump.
Totem jumping gives an extra 20% to the applied jump velocity.

throwVector was changed from [1f,1f] to [1f, 0.7f]. throwPower was
changed from 6f to 8f.