Add check for nested totems Modified various methods in StackTotem
StackTotemInPos now checks the newly picked up totem for nested totems.
If there are nested totems, the nested tokens are removed from the
picked up totem's listStackedTotems and are put into the current totems
listStackedTotems.

TransferListToTargetTotem has been modified to have a donor and
recipient StackTotem inputs. This is so that the nester totem check
can take totems from the correct StackTotem class instead of always
specifying itself as the target.

RemoveTotemFromStackList has been modified to specify a donor stack.
This is so that totems are removed from a specified stack instead of
itself.