Switch scenes 2 and 3 around Modify finishLevelTriggers for scenes 2 and 3 Modified level title text Modifed credits text
Switched levels 2 and 3 around so that control introductions are more
logical. As levels 2 and 3 were switched around, the finishLevelTrigger
for levels 2 and 3 were also modified to go to the correct levels.

Level title text was moved to be in the top left corner of the screen.
The text size was also increased to 35.

The credits text in the title menu was changed to size 35 as well.